services/detector-py/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
schedule
aiohttp
sentence-transformers
scikit-learn
web3
